Move graph component into its own file
